feat: 세션 기반 인증/권한 처리(HttpSession) 적용(BE)

+++ clean4u/src/main/java/org/example/clean4u/laundryOption/LaundryOptionController.java
@@ -1,7 +1,10 @@
 package org.example.clean4u.laundryOption;
 
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.example.clean4u._core.exception.Exception401;
+import org.example.clean4u.employee.Employee;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,7 +21,11 @@ public class LaundryOptionController {
 
     // http://localhost:8080/laundry-option
     @GetMapping("/laundry-option")
-    public String laundryOptionList(Model model) {
+    public String laundryOptionList(Model model, HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         List<LaundryOption> laundryOptionList = repository.findAll();
         model.addAttribute("laundryOptionList", laundryOptionList);
         return "/laundryOption/list-form";
@@ -26,7 +33,11 @@ public class LaundryOptionController {
 
     // http://localhost:8080/laundry-option/{id}
     @GetMapping("/laundry-option/{id}")
-    public String detail(@PathVariable Long id, Model model) {
+    public String detail(@PathVariable Long id, Model model, HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         LaundryOption laundryOption = repository.findById(id);
         model.addAttribute("laundryOption", laundryOption);
 
@@ -35,13 +46,21 @@ public class LaundryOptionController {
 
     // http://localhost:8080/laundry-option/save
     @GetMapping("/laundry-option/save")
-    public String saveForm() {
+    public String saveForm(HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         return "laundryOption/save-form";
     }
 
     // http://localhost:8080/laundry-option/save
     @PostMapping("/laundry-option/save")
-    public String saveProc(@Valid LaundryOptionRequest.SaveDTO saveDTO) {
+    public String saveProc(@Valid LaundryOptionRequest.SaveDTO saveDTO, HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         LaundryOption laundryOption = saveDTO.toEntity();
         repository.save(laundryOption);
         return "redirect:/laundry-option";
@@ -49,7 +68,11 @@ public class LaundryOptionController {
 
     // http://localhost:8080/laundry-option/{id}/update
     @GetMapping("/laundry-option/{id}/update")
-    public String updateForm(@PathVariable Long id, Model model) {
+    public String updateForm(@PathVariable Long id, Model model, HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         LaundryOption laundryOption = repository.findById(id);
         model.addAttribute("laundryOption", laundryOption);
         return "laundryOption/update-form";
@@ -57,14 +80,22 @@ public class LaundryOptionController {
 
     // http://localhost:8080/laundry-option/{id}/update
     @PostMapping("/laundry-option/{id}/update")
-    public String updateProc(@PathVariable Long id, @Valid LaundryOptionRequest.UpdateDTO updateDTO) {
+    public String updateProc(@PathVariable Long id, @Valid LaundryOptionRequest.UpdateDTO updateDTO, HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         repository.updateById(id, updateDTO);
         return "redirect:/laundry-option/{id}";
     }
 
     // http://localhost:8080/laundry-option/{id}/delete
     @PostMapping("/laundry-option/{id}/delete")
-    public String delete(@PathVariable Long id) {
+    public String delete(@PathVariable Long id, HttpSession session) {
+        Employee sessionUser = (Employee) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new Exception401("로그인이 필요합니다.");
+        }
         repository.deleteById(id);
         return "redirect:/laundry-option";
     }
